overlay: cache metrics so we don't keep looking them up.
